删除上游源类型,以及删除组织机构树类型的逻辑补充

+++ src/main/java/com/qtgl/iga/service/impl/DeptTreeTypeServiceImpl.java
@@ -1,7 +1,9 @@
 package com.qtgl.iga.service.impl;
 
 import com.qtgl.iga.bo.DeptTreeType;
+import com.qtgl.iga.bo.Node;
 import com.qtgl.iga.dao.DeptTreeTypeDao;
+import com.qtgl.iga.dao.NodeDao;
 import com.qtgl.iga.service.DeptTreeTypeService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,8 @@ public class DeptTreeTypeServiceImpl implements DeptTreeTypeService {
 
     @Autowired
     DeptTreeTypeDao deptTreeTypeDao;
+    @Autowired
+    NodeDao nodeDao;
 
     @Override
     public List<DeptTreeType> findAll(Map<String, Object> arguments, String domain) {
@@ -24,6 +28,11 @@ public class DeptTreeTypeServiceImpl implements DeptTreeTypeService {
 
     @Override
     public DeptTreeType deleteDeptTreeType(Map<String, Object> arguments, String domain) throws Exception {
+        //查询是否有绑定的node
+        List<Node> nodeList = nodeDao.findByTreeTypeId((String)arguments.get("id"));
+        if(null != nodeList && nodeList.size()>0){
+            throw new Exception("删除组织机构树类型失败,有绑定的node,请查看后再删除");
+        }
         return deptTreeTypeDao.deleteDeptTreeType(arguments, domain);
     }
