+++ src/main/java/com/qtgl/iga/service/impl/UpstreamTypeServiceImpl.java
@@ -1,8 +1,10 @@
 package com.qtgl.iga.service.impl;
 
 
+import com.qtgl.iga.bo.NodeRules;
 import com.qtgl.iga.bo.UpstreamType;
 import com.qtgl.iga.bo.UpstreamTypeField;
+import com.qtgl.iga.dao.NodeRulesDao;
 import com.qtgl.iga.dao.UpstreamTypeDao;
 import com.qtgl.iga.service.UpstreamTypeService;
 import com.qtgl.iga.vo.UpstreamTypeVo;
@@ -10,6 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.print.DocFlavor;
 import java.util.List;
 import java.util.Map;
 
@@ -19,6 +22,8 @@ public class UpstreamTypeServiceImpl implements UpstreamTypeService {
 
     @Autowired
     UpstreamTypeDao upstreamTypeDao;
+    @Autowired
+    NodeRulesDao nodeRulesDao;
 
     @Override
     public List<UpstreamTypeVo> findAll(Map<String, Object> arguments, String domain) {
@@ -27,6 +32,11 @@ public class UpstreamTypeServiceImpl implements UpstreamTypeService {
 
     @Override
     public UpstreamType deleteUpstreamType(Map<String, Object> arguments, String domain) throws Exception {
+        //查看是否有关联node_rules
+        List<NodeRules> nodeRules = nodeRulesDao.findNodeRulesByUpStreamTypeId((String) arguments.get("id"));
+        if(null != nodeRules && nodeRules.size()>0){
+            throw new Exception("删除上游源类型失败,有绑定的node规则,请查看后再删除");
+        }
         return upstreamTypeDao.deleteUpstreamType(arguments, domain);
     }
